+++ docs/source/releasenotes.rst
@@ -4,6 +4,9 @@ Release notes
 `Upcoming release <https://github.com/robocorp/rpaframework/projects/3#column-16713994>`_
 +++++++++++++++++
 
+- Library **RPA.Email.ImapSmtp**: Keyword ``Email To Document`` for converting HTML or
+  Text e-mails into Word documents (:issue:`295`)
+
 `Released <https://pypi.org/project/rpaframework/#history>`_
 +++++++++