+++ docs/source/releasenotes.rst
@@ -6,6 +6,10 @@ Release notes
 
 - Library **RPA.Email.ImapSmtp**: Keyword ``Email To Document`` for converting HTML or
   Text e-mails into Word documents (:issue:`295`)
+- Library **RPA.Robocorp.WorkItems**:
+  - Removed Keyword ``Parse Work Item From Email``
+  - Automatically loads e-mail body formats like JSON/YAML/Text/HTML into "parsedEmail"
+    work item variable
 
 `Released <https://pypi.org/project/rpaframework/#history>`_
 +++++++++